Fix(pricing): make the cards line up at the bottom
The alignment was being managed by a global rule for the .inner class,
which resulted in the first card having a higher bottom line than the
other cards.

This could be perceived as a mistake in the styling, so I've fixed that
up by giving the pricing page it's own pricing-inner class and removing
the alignment that caused it to break.